adding update post and it's working

+++ app/main/views.py
@@ -99,14 +99,20 @@ def deleteComment(comment_id):
     flash('Your comment has been deleted!', 'success')
     return redirect(url_for('main.blog'))
 
+
+
 @main.route('/delete/post/<int:post_id>',methods=['POST'])
 @login_required
 def deletePost(post_id):
-    post = Post.query.filter(Post.id == post_id).first()
+    post = Post.query.get_or_404(post_id)
+    if post.author != current_user:
+        abort(403)
     db.session.delete(post)
     db.session.commit()
-    flash('Your post has been deleted!', 'success')
-    return redirect(url_for('main.blog'))
+    flash('Post has been deleted!', 'success')
+    return redirect(url_for('main.post'))
+
+
 
 @main.route("/post/<int:post_id>/update", methods=['GET', 'POST'])
 @login_required
@@ -120,10 +126,10 @@ def updatePost(post_id):
         post.content = form.content.data
         db.session.commit()
         flash('Post has been updated!', 'success')
-        return redirect(url_for('post', post_id=post.id))
+        return redirect(url_for('main.blog', post_id=post.id))
     elif request.method == 'GET':
         form.title.data = post.title
         form.content.data = post.content
-    return render_template('create_post.html', title='Update Post',
+    return render_template('post.html', title='Update Post',
                            form=form, legend='Update Post')
